+++ app/src/components/homePageMainContent/homePageMainContent.tsx
@@ -1,13 +1,14 @@
 import React from "react";
 import FeaturedContentComponent from "../featuredContent/featuredContent";
-import SwimLanes from "../swimlanes/swimLanes";
-import SwimLanesLoading from "../swimlanes/swimLanesLoading";
 import { useNumColumns } from "../../hooks/useNumColumns";
 import { useEffect, useState } from "react";
+import CollectionLanes from "../collectionLanes/collectionLanes";
+import CollectionLanesLoading from "../collectionLanes/collectionLanesLoading";
 
 const HomePageMainContent = ({ data }) => {
   const numColumns = useNumColumns();
   const [isLoaded, setIsLoaded] = useState(false);
+  console.log(data.lanesWithNumItems);
 
   useEffect(() => {
     setIsLoaded(true);
@@ -15,19 +16,22 @@ const HomePageMainContent = ({ data }) => {
 
   return isLoaded ? (
     <>
-      <SwimLanes
+      <CollectionLanes
         numColumns={numColumns}
         lanesWithNumItems={[data.lanesWithNumItems[0]]}
+        isSwimLane={true}
       />
       <FeaturedContentComponent randomNumber={data.randomNumber} />
-      <SwimLanes
+
+      <CollectionLanes
         numColumns={numColumns}
         lanesWithNumItems={data.lanesWithNumItems.slice(1)}
+        isSwimLane={true}
       />
     </>
   ) : (
     <>
-      <SwimLanesLoading />
+      <CollectionLanesLoading />
     </>
   );
 };