+++ app/src/models/collectionCard.ts
@@ -8,16 +8,16 @@ export class CollectionCardModel {
   url: string;
   imageID: string | null;
   imageURL: string;
-  numItems: number;
+  numberOfDigitizedItems: number;
   containsOnSiteMaterials: boolean;
 
   constructor(data: any) {
     this.uuid = data.uuid;
     this.title = data.title;
     this.url = data.url;
-    this.imageID = data.image_id;
-    this.imageURL = imageURL(data.image_id, "full", "288,", "0");
-    this.numItems = data.numItems || 0;
+    this.imageID = data.imageID;
+    this.imageURL = imageURL(data.imageID, "full", "288,", "0");
+    this.numberOfDigitizedItems = data.numberOfDigitizedItems || 0;
     this.containsOnSiteMaterials = data.containsOnSiteMaterials || false;
   }
 }